+++ mealpy/utils/validator.py
@@ -5,8 +5,31 @@
 # --------------------------------------------------%
 
 import numpy as np
-from mealpy.utils.boundary import is_in_bound, is_str_in_list
+import operator
 from mealpy.utils.logger import Logger
+
+
+def is_in_bound(value, bound):
+    ops = None
+    if type(bound) is tuple:
+        ops = operator.lt
+    elif type(bound) is list:
+        ops = operator.le
+    if bound[0] == float("-inf") and bound[1] == float("inf"):
+        return True
+    elif bound[0] == float("-inf") and ops(value, bound[1]):
+        return True
+    elif ops(bound[0], value) and bound[1] == float("inf"):
+        return True
+    elif ops(bound[0], value) and ops(value, bound[1]):
+        return True
+    return False
+
+
+def is_str_in_list(value: str, my_list: list):
+    if type(value) == str and my_list is not None:
+        return True if value in my_list else False
+    return False
 
 
 class Validator: